drivers: Implement semihosting based console driver for log
Semihosting is a mechanism that enables target to communicate and use
I/O facilities on a host computer which is running a debugger, such as
GDB. The I/O facilities include character {read|write} {from|to} the
semihosting host side console or a file. In other words, OP-TEE OS can
output log to the host side console or the host side file, if there is a
semihosting host and OP-TEE OS requests the semihosting operations to
that host.

This commit implements a simple console driver which uses semihosting
operations to read/write the character. There are two paths to output
the log:
- If the caller of semihosting_console_init() provides the path of the
  file, the driver will try to open that file, and output the log to
  that host side file.
- If the caller of semihosting_console_init() does not provide the path
  of the file, the driver will try to output the log to host side
  console directly.

Signed-off-by: Alvin Chang <alvinga@andestech.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>